Using DTO projections to get CollectionEventInfoDTO

+++ src/main/java/edu/ualberta/med/biobank/dtos/CollectionEventSummaryDTO.java
@@ -5,8 +5,8 @@ import java.util.Date;
 public record CollectionEventSummaryDTO(
     Integer id,
     Integer visitNumber,
-    Integer specimenCount,
-    Integer aliquotCount,
+    Long specimenCount,
+    Long aliquotCount,
     Date createdAt,
     String status
 ) {}
